requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets>=12.0
python-dotenv==1.0.1
pydantic==2.10.3

# Database
neo4j==5.27.0
graphiti-core==0.3.17

# LLM
litellm

# MCP
fastmcp

# Utilities
httpx
